# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_opsmx_ai-guardian-api_pkg_repository.ProjectWithLatestScan:
    properties:
      branch:
        type: string
      id:
        type: string
      name:
        type: string
      repo:
        type: string
    type: object
...
  github_com_opsmx_ai-guardian-api_pkg_service.CreateProjectRequest:
    properties:
      branch:
        type: string
      hub_id:
        type: string
      integration_id:
        type: string
      name:
        type: string
      repoName:
        type: string
      schedule_time:
        type: integer
      scheduled_scan:
        type: boolean
    type: object
  github_com_opsmx_ai-guardian-api_pkg_service.ProjectsListAllResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_opsmx_ai-guardian-api_pkg_repository.ProjectWithLatestScan'
        type: array
      total:
        type: integer
    type: object
...
  github_com_opsmx_ai-guardian-api_pkg_service.UpdateProjectRequest:
    properties:
      branch:
        type: string
      hub_id:
        type: string
      id:
        type: string
      integration_id:
        type: string
      name:
        type: string
      repoName:
        type: string
      schedule_time:
        type: integer
      scheduled_scan:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: AI Guardian API Service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: AI Guardian API
  version: "1.0"
paths:
...
  /api/v1/hubs/{hub_id}/projects/summaries:
    get:
      consumes:
      - application/json
      description: Returns a list of project summaries for the specified hub
      parameters:
      - description: Hub ID
        in: path
        name: hub_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid hub ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get project summaries for hub
      tags:
      - Projects
...
  /api/v1/projects:
    post:
      consumes:
      - application/json
      description: Creates a new project with the provided details
      parameters:
      - description: Project creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_opsmx_ai-guardian-api_pkg_service.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new project
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Updates a existing project with the provided details
      parameters:
      - description: Project update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_opsmx_ai-guardian-api_pkg_service.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a existing project
      tags:
      - Projects
  /api/v1/projects/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes the project with the specified ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Comma-separated team IDs (optional)
        in: query
        name: teamIds
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid project ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Project not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete project by ID
      tags:
      - Projects
    get:
      consumes:
      - application/json
      description: Returns the project with the specified ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid project ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Project not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get project by ID
      tags:
      - Projects
  /api/v1/projects/{id}/stats:
    get:
      consumes:
      - application/json
      description: Returns the project stats with the specified ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Database selector (optional)
        in: query
        name: db
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid project ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Project not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get project stats by ID
      tags:
      - Projects
  /api/v1/projects/list/all:
    get:
      consumes:
      - application/json
      description: Returns all projects with repo/branch taken from each project's
        latest scan (scans.created_at DESC)
      parameters:
      - description: Hub ID
        in: query
        name: hubId
        required: true
        type: string
      - description: Optional project name search (ILIKE)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_opsmx_ai-guardian-api_pkg_service.ProjectsListAllResponse'
        "400":
          description: Missing/invalid hubId
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: List all projects with latest repo/branch
      tags:
      - Projects
  /api/v1/projects/summaries/count:
    get:
      consumes:
      - application/json
      description: Returns the count of project summaries based on filters
      parameters:
      - description: Filter by team ID
        in: query
        name: team_id
        type: string
      - description: Filter by hub ID
        in: query
        name: hub_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get project summary count
      tags:
      - Projects
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"